+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] ##
+### Fixed
+- Fixed a bug where if the 'y' is capitalized when confirming a frame delete, it is treated as a no.
 
 ## [1.0.1] - 2018-12-26 ##
 ### Added