+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] ##
+## [1.1.0] - 2019-02-01 ##
 ### Added
 - Added 'now' keyword to '--end' flag time string. Gets replaced with current time and date.
 - Added '-d' flag to 'log' command that shows frames just for yesterday
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Core commands implemented
 - Data storage format stabilized
 
-[Unreleased]: https://github.com/gochrono/chrono/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/gochrono/chrono/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/gochrono/chrono/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/gochrono/chrono/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/gochrono/chrono/compare/v1.0.0-beta...v1.0.0